+++ views/layouts/head.php
@@ -1,0 +1,28 @@
+<?php header('Content-Type: text/html; charset=UTF-8');
+/**
+ * Created by PhpStorm.
+ * User: fryntiz
+ * Date: 7/10/18
+ * Time: 1:39
+ */
+/**
+ * @author    Raúl Caro Pastorino
+ * @link      https://fryntiz.es
+ * @copyright Copyright (c) 2018 Raúl Caro Pastorino
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html
+ **/
+
+use yii\helpers\Html;
+?>
+
+<!DOCTYPE html>
+<html lang="<?= Yii::$app->language ?>">
+<head>
+    <meta charset="<?= Yii::$app->charset ?>">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <?= Html::csrfMetaTags() ?>
+    <title><?= Html::encode($this->title) ?></title>
+    <?php $this->head() ?>
+</head>
+<body>
