+++ views/layouts/main.php
@@ -1,12 +1,22 @@
-<?php
+<?php header('Content-Type: text/html; charset=UTF-8');
+/**
+ * Created by PhpStorm.
+ * User: fryntiz
+ * Date: 7/10/18
+ * Time: 1:39
+ */
+/**
+ * @author    Raúl Caro Pastorino
+ * @link      https://fryntiz.es
+ * @copyright Copyright (c) 2018 Raúl Caro Pastorino
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html
+ **/
 
 /* @var $this \yii\web\View */
 /* @var $content string */
 
 use app\widgets\Alert;
 use yii\helpers\Html;
-use yii\bootstrap\Nav;
-use yii\bootstrap\NavBar;
 use yii\widgets\Breadcrumbs;
 use app\assets\AppAsset;
 
@@ -15,54 +25,13 @@ AppAsset::register($this);
 $params = Yii::$app->params;
 
 ?>
+
 <?php $this->beginPage() ?>
-<!DOCTYPE html>
-<html lang="<?= Yii::$app->language ?>">
-<head>
-    <meta charset="<?= Yii::$app->charset ?>">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <?= Html::csrfMetaTags() ?>
-    <title><?= Html::encode($this->title) ?></title>
-    <?php $this->head() ?>
-</head>
-<body>
+<?php require_once 'head.php' ?>
 <?php $this->beginBody() ?>
 
 <div class="wrap">
-    <?php
-    NavBar::begin([
-        'brandLabel' => Yii::$app->name,
-        'brandUrl' => Yii::$app->homeUrl,
-        'options' => [
-            'class' => 'navbar-inverse navbar-fixed-top',
-        ],
-    ]);
-    echo Nav::widget([
-        'options' => ['class' => 'navbar-nav navbar-right'],
-        'items' => [
-            ['label' => \Yii::t('main', 'home'), 'url' => ['/site/index']],
-            ['label' => \Yii::t('main', 'users'), 'url' => ['/usuarios/index']],
-            ['label' => \Yii::t('main', 'torrents'), 'url' => ['/torrents/index']],
-            ['label' => \Yii::t('main', 'about'), 'url' => ['/site/about']],
-            ['label' => \Yii::t('main', 'contact'), 'url' => ['/site/contact']],
-            Yii::$app->user->isGuest ? (
-            ['label' => \Yii::t('main', 'login'), 'url' => ['/site/login']]
-            ) : (
-                '<li>'
-                . Html::beginForm(['/site/logout'], 'post')
-                . Html::submitButton(
-                    \Yii::t('main', 'home').'(' .
-                    Yii::$app->user->identity->nick . ')',
-                    ['class' => 'btn btn-link logout']
-                )
-                . Html::endForm()
-                . '</li>'
-            )
-        ],
-    ]);
-    NavBar::end();
-    ?>
+    <?php require_once 'navbar.php' ?>
 
     <div class="container">
         <?= Breadcrumbs::widget([
@@ -73,16 +42,7 @@ $params = Yii::$app->params;
     </div>
 </div>
 
-<footer class="footer">
-    <div class="container">
-        <p class="pull-left">&copy; <?= $params['sitename']; ?> 2018-<?= date
-            ('Y')
-            ?></p>
-
-        <p class="pull-right"><a href="http://www.fryntiz.es" alt="web de
-        Raúl Caro Pastorino">Raúl Caro Pastorino</a></p>
-    </div>
-</footer>
+<?php require_once 'footer.php' ?>
 
 <?php $this->endBody() ?>
 </body>
