+++ views/site/index.php
@@ -1,4 +1,16 @@
-<?php
+<?php header('Content-Type: text/html; charset=UTF-8');
+/**
+ * Created by PhpStorm.
+ * User: fryntiz
+ * Date: 7/10/18
+ * Time: 1:40
+ */
+/**
+ * @author    Raúl Caro Pastorino
+ * @link      https://fryntiz.es
+ * @copyright Copyright (c) 2018 Raúl Caro Pastorino
+ * @license   https://www.gnu.org/licenses/gpl-3.0-standalone.html
+ **/
 
 /* @var $this yii\web\View */
